+++ plugins/common-repo/src/main/groovy/com/apgsga/common/repo/plugin/ApgCommonRepoPlugin.groovy
@@ -1,6 +1,6 @@
 package com.apgsga.common.repo.plugin
 
-import com.apgsga.common.repo.extensions.LoaderManager
+
 import groovy.transform.CompileStatic
 import org.gradle.api.Plugin
 import org.gradle.api.Project
